Refact : 코드 수정
