+++ src/main/java/com/example/catchroom_be/Test/repository/TestEntityRepository.java
@@ -1,6 +1,6 @@
-package com.example.catchroom_be.Test.Repository;
+package com.example.catchroom_be.Test.repository;
 
-import com.example.catchroom_be.Test.Entity.TestEntity;
+import com.example.catchroom_be.Test.entity.TestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TestEntityRepository extends JpaRepository<TestEntity, Long> {
