--- src/main/java/com/example/catchroom_be/User/Controller/OAuthController.java
@@ -1,16 +1,0 @@
-package com.example.catchroom_be.User.Controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-public class OAuthController {
-
-    @PostMapping("/oauth2/callback")
-    public String getAuthCode(@RequestParam String authCode) {
-        return authCode;
-    }
-}
-
