+++ src/main/java/com/example/catchroom_be/User/dto/OauthAuthCodeRequest.java
@@ -1,0 +1,9 @@
+package com.example.catchroom_be.User.dto;
+
+public class OauthAuthCodeRequest {
+    String authCode;
+
+    public String getAuthCode() {
+        return authCode;
+    }
+}
